refactor components with add edit ui
