--- src/TodoList.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function TodoList() {
-  return <div>TodoList</div>;
-}
-
-export default TodoList;
